requirements.txt
telethon
python-dotenv
apscheduler
aiohttp
psutil
colorlog
python-json-logger